関数名とコメントの修正AutoMoverBaseの設計修正
