+++ Assets/MyAssets/Script/AutoTileGroupMover.cs
@@ -19,30 +19,21 @@ public class AutoTileGroupMover : AutoMoverBase
     Action<GameObject> repositionAction;
 
     /// <summary>
+    /// 初期化処理
+    /// </summary>
+    void Start()
+    {
+        base.Initialize();
+        base.SetScreenChangerAction(ResetImage);
+    }
+
+    /// <summary>
     /// Actionに関数を登録する処理
     /// </summary>
     /// <param name="action">セットするAction</param>
     public void SetAction(Action<GameObject> action)
     {
         repositionAction = action;
-    }
-
-    /// <summary>
-    /// 初期化処理
-    /// リセット関数の登録と初期ポジションの保存
-    /// </summary>
-    void Start()
-    {
-        gameToResultScreenChanger.SetAction(Initialize);
-        keepPosition = myRectTransform.position;
-    }
-
-    /// <summary>
-    /// 更新処理
-    /// </summary>
-    void Update()
-    {
-        base.AutoMove();
     }
 
     /// <summary>
@@ -72,24 +63,5 @@ public class AutoTileGroupMover : AutoMoverBase
         {
             tile.SetActive(false);
         }
-    }
-
-    /// <summary>
-    /// ループ中に行う初期化処理
-    /// </summary>
-    void Initialize()
-    {
-        foreach (var tile in tileObjects)
-        {
-            tile.SetActive(true);
-        }
-
-        foreach (var tile in breakTileObjects)
-        {
-            tile.SetActive(false);
-        }
-        isMove = false;
-        receivedMoveSpeed = 0.0f;
-        myRectTransform.position = keepPosition;
     }
 }
